+++ app/src/androidTest/java/woowacourse/movie/view/MovieDetailActivityTest.kt
@@ -14,7 +14,7 @@ import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import woowacourse.movie.R
-import woowacourse.movie.detail.view.MovieDetailActivity
+import woowacourse.movie.detail.MovieDetailActivity
 
 @RunWith(AndroidJUnit4::class)
 class MovieDetailActivityTest {
