+++ app/src/androidTest/java/woowacourse/movie/view/MovieMainActivityTest.kt
@@ -10,7 +10,7 @@ import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import woowacourse.movie.R
-import woowacourse.movie.main.view.MovieMainActivity
+import woowacourse.movie.main.MovieMainActivity
 
 @RunWith(AndroidJUnit4::class)
 class MovieMainActivityTest {
