+++ app/src/androidTest/java/woowacourse/movie/view/MovieResultActivityTest.kt
@@ -12,7 +12,7 @@ import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import woowacourse.movie.R
-import woowacourse.movie.result.view.MovieResultActivity
+import woowacourse.movie.result.MovieResultActivity
 
 @RunWith(AndroidJUnit4::class)
 class MovieResultActivityTest {
