+++ app/src/main/AndroidManifest.xml
@@ -13,15 +13,15 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".result.view.MovieResultActivity"
+            android:name=".result.MovieResultActivity"
             android:exported="false"
-            android:parentActivityName=".detail.view.MovieDetailActivity"/>
+            android:parentActivityName=".detail.MovieDetailActivity"/>
         <activity
-            android:name=".detail.view.MovieDetailActivity"
+            android:name=".detail.MovieDetailActivity"
             android:exported="false"
-            android:parentActivityName=".main.view.MovieMainActivity"/>
+            android:parentActivityName=".main.MovieMainActivity"/>
         <activity
-            android:name=".main.view.MovieMainActivity"
+            android:name=".main.MovieMainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
