+++ app/src/main/res/layout/activity_movie_result.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".result.view.MovieResultActivity">
+    tools:context=".result.MovieResultActivity">
 
 
     <TextView
